Moved more date/time-related utility functions to the other class.
